unit_tests/requirements.txt
-r ../requirements-base.txt

requests==2.32.5
pytest-mock==3.15.1
pytest-cov==7.0.0
responses==0.25.8